feat(CastList): display error message when cast data fails to load
This commit adds error handling to the CastList component. If the useGetCast hook encounters an error, an error message is now displayed to the user, informing them that the cast data could not be retrieved. This improves the user experience by providing feedback when there are issues loading cast information.